requirements.txt
Django==1.5.4
MySQL-python==1.2.4
South==0.8.2
django-admin-tools==0.5.1
django-ajax-validation==0.1.3
django-email-confirmation==0.2
django-tagging==0.3.1
gunicorn==18.0
pytils==0.2.3
pytz==2013.7
setproctitle==1.1.7
wsgiref==0.1.2
